use hashable_ord on short_ids
